Newsletter: move getNewsletterContentData out of newsletter.ts (fix CI)
The unit tests import the pure date/data helpers from newsletter.ts, but that
module had gained a top-level `import { getImage } from "astro:assets"` for the
image cropping. Playwright's plain Node ESM loader can't resolve the `astro:`
virtual scheme, so `playwright test` failed to even load the suite.

Move getNewsletterContentData (the only astro:assets consumer) into a new
utils/newsletter-content.ts, leaving newsletter.ts importable from plain Node.
Repoint the five route consumers. No behaviour change; getImage stays a static
import (in the new file). Export PipelineWorkflow for the moved signature.
