+++ sites/main-site/src/pages/newsletter/[year]/[month]/simple.astro
@@ -1,7 +1,8 @@
 ---
 import { getCollection } from "astro:content";
 import pipelines_json from "@public/pipelines.json";
-import { getNewsletterContentData, getNewsletterStaticPaths } from "@utils/newsletter";
+import { getNewsletterStaticPaths } from "@utils/newsletter";
+import { getNewsletterContentData } from "@utils/newsletter-content";
 import NewsletterSimpleLayout from "@layouts/NewsletterSimpleLayout.astro";
 
 const images = import.meta.glob("/src/assets/**");
