+++ sites/main-site/utils/newsletter.ts
@@ -1,12 +1,11 @@
 import { octokit } from "@components/octokit.js";
-import { getImage } from "astro:assets";
 import { addMonths } from "date-fns";
 
 // ========================================
 // TYPE DEFINITIONS
 // ========================================
 
-interface PipelineWorkflow {
+export interface PipelineWorkflow {
     name: string;
     description: string;
     created_at: string;
@@ -460,132 +459,4 @@ export async function getNewsletterStaticPaths(
             nextMonth: months[index - 1] || null,
         },
     }));
-}
-
-/**
- * Shared content data fetching for a single newsletter month.
- * Returns the data object expected by NewsletterContent component.
- */
-export async function getNewsletterContentData(
-    getCollectionFn: (name: string) => Promise<any[]>,
-    pipelines: PipelineWorkflow[],
-    year: number,
-    month: number,
-    allProposals: RawProposal[],
-    images: Record<string, () => Promise<any>>,
-) {
-    const monthName = getMonthName(month);
-    const baseUrl = "https://nf-co.re";
-
-    // Content month (previous calendar month — newsletter looks back)
-    const contentMonth = month === 1 ? 12 : month - 1;
-    const contentYear = month === 1 ? year - 1 : year;
-
-    let blogPosts = await getCollectionFn("blog");
-    blogPosts = blogPosts.filter((post: any) => new Date(post.data.pubDate) < new Date());
-
-    let events = await getCollectionFn("events");
-    events = events.filter((e: any) => e.id.split("/").length === 2);
-
-    const thisMonthBlogPosts = getBlogPostsForMonth(blogPosts, contentYear, contentMonth);
-    const olderBlogPosts = getBlogPostsForPreviousMonths(blogPosts, contentYear, contentMonth, 2);
-    const releases = getPipelineReleasesForMonth(pipelines, contentYear, contentMonth);
-    const firstReleases = releases.filter((r) => r.isFirstRelease);
-    const otherReleases = releases.filter((r) => !r.isFirstRelease);
-    const newPipelines = getNewPipelines(pipelines, contentYear, contentMonth);
-    const recentEvents = getEventsForMonth(events, contentYear, contentMonth);
-    const allAdvisories = await getCollectionFn("advisories");
-    const monthAdvisories = getAdvisoriesForMonth(allAdvisories, contentYear, contentMonth);
-    const olderAdvisories = getAdvisoriesForPreviousMonths(allAdvisories, contentYear, contentMonth, 2);
-    const proposals = getProposalsForMonth(allProposals, contentYear, contentMonth);
-    const pipelineProposals = proposals.filter((p) => p.category === "pipeline");
-    const otherProposals = proposals.filter((p) => p.category === "other");
-
-    const upcomingEvents = [...getEventsForMonth(events, year, month), ...getFutureEvents(events, year, month, 1)];
-
-    // Build email preview summary (~90 chars for Gmail)
-    const previewParts: string[] = [];
-    if (thisMonthBlogPosts.length > 0)
-        previewParts.push(`${thisMonthBlogPosts.length} blog post${thisMonthBlogPosts.length > 1 ? "s" : ""}`);
-    if (monthAdvisories.length > 0)
-        previewParts.push(`${monthAdvisories.length} advisor${monthAdvisories.length > 1 ? "ies" : "y"}`);
-    if (releases.length > 0) previewParts.push(`${releases.length} pipeline release${releases.length > 1 ? "s" : ""}`);
-    if (upcomingEvents.length > 0)
-        previewParts.push(`${upcomingEvents.length} upcoming event${upcomingEvents.length > 1 ? "s" : ""}`);
-    const previewText =
-        previewParts.length > 0 ? previewParts.join(", ") + "." : `Community news for ${monthName} ${year}.`;
-    const newsletterUrl = `${baseUrl}/newsletter/${year}/${String(month).padStart(2, "0")}`;
-
-    // Pre-resolve all image sources to a cropped 16:9 thumbnail. Header images
-    // come in any aspect ratio, and email clients ignore `object-fit`, so we crop
-    // the actual file here (works for both bundled /assets images and whitelisted
-    // remote URLs). Falls back to the original src if optimisation fails (e.g. a
-    // non-whitelisted domain), which still renders — just uncropped.
-    async function resolveImageSrc(headerImage: string): Promise<string> {
-        // `/assets/…` blog/event images are bundled (src/assets); resolve them to
-        // their bundled asset. Anything not in the bundle (e.g. a /public path) or a
-        // remote URL is used as-is.
-        let src: any = headerImage;
-        if (headerImage.startsWith("/assets/")) {
-            const loader = images["/src" + headerImage];
-            if (!loader) return headerImage;
-            src = ((await loader()) as any).default;
-        }
-        try {
-            // JPEG (not WebP) for broad email-client support, incl. Outlook on Windows.
-            const thumb = await getImage({
-                src,
-                width: 480,
-                height: 270,
-                fit: "cover",
-                position: "center",
-                format: "jpeg",
-            });
-            return thumb.src;
-        } catch {
-            // If optimisation fails, fall back to the *un-cropped* image that still
-            // loads — the bundled asset src for local images, the URL for remote.
-            return typeof src === "string" ? src : src.src;
-        }
-    }
-
-    const blogImageSrcs = new Map<string, any>(
-        await Promise.all(
-            [...thisMonthBlogPosts, ...olderBlogPosts].map(
-                async (post) => [post.id, await resolveImageSrc((post as any).data.headerImage)] as [string, any],
-            ),
-        ),
-    );
-    const eventImageSrcs = new Map<string, any>(
-        await Promise.all(
-            upcomingEvents
-                .filter((event) => (event as any).data.headerImage)
-                .map(
-                    async (event) =>
-                        [event.id, await resolveImageSrc((event as any).data.headerImage)] as [string, any],
-                ),
-        ),
-    );
-
-    return {
-        year,
-        month,
-        monthName,
-        baseUrl,
-        newsletterUrl,
-        previewText,
-        thisMonthBlogPosts,
-        olderBlogPosts,
-        firstReleases,
-        otherReleases,
-        newPipelines,
-        recentEvents,
-        upcomingEvents,
-        pipelineProposals,
-        otherProposals,
-        monthAdvisories,
-        olderAdvisories,
-        blogImageSrcs,
-        eventImageSrcs,
-    };
 }
